pyproject.toml
[project]
name = "langraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9.18"
dependencies = [
    "langchain[anthropic]>=0.3.25",
    "langgraph>=0.4.3",
    "python-dotenv>=1.1.0",
]
